Update CSR list
This change replaces the list of CSR names with the latest version in the
https://github.com/riscv/riscv-opcodes.git repository. The list was generated
by adding the five-line function to "parse_opcodes" (a Python script in the
riscv-opcodes project), arranging for this function to be called, and then
piping the output through sort.

[BEGIN PYTHON]
def make_csr():
  for (n,s) in csrs:
    print('    case 0x%04x: return "%s";' % (n,s))
  for (n,s) in csrs32:
    print('    case 0x%04x: return "%s";' % (n,s))
[END PYTHON]

Of note is that the original version had an error whereby the strings for
"pmpaddr13" and "pmpaddr14" were swapped.